feat: extract landing nav controls
