+++ client/src/pages/Home.jsx
@@ -4,32 +4,13 @@
  * @since 2025-11-05
  * @purpose Temporary full-screen landing with brand logo and primary CTAs.
  */
-import { useState } from 'react';
 import { Link } from 'react-router-dom';
-import { PRIMARY_NAV_LINKS } from '../utils/navigation';
+import LandingNavControls from '../components/LandingNavControls.jsx';
 
 export default function Home() {
-  const [menuOpen, setMenuOpen] = useState(false);
-
-  const toggleMenu = () => setMenuOpen((v) => !v);
-  const closeMenu = () => setMenuOpen(false);
-
   return (
     <div className="landing-basic">
-      <header className="landing-basic__chrome">
-        <button
-          type="button"
-          className="landing-basic__menu"
-          aria-label="Open navigation"
-          aria-controls="landing-drawer"
-          aria-expanded={menuOpen}
-          onClick={toggleMenu}
-        >
-          <span />
-          <span />
-          <span />
-        </button>
-      </header>
+      <LandingNavControls />
 
       <main className="landing-basic__body">
         <img
@@ -55,28 +36,6 @@ export default function Home() {
         COMPLETED: Routed play button to /sessions (active sessions page). Completed by Kelly Burden - Nov 2025*/}
       </main>
 
-      {/* Simple slide-out drawer for quick navigation while on the landing view. */}
-      {/* TODO (Frontend): read links from PRIMARY_NAV_LINKS to avoid duplicating routes here. 
-      COMPLETED: Implemented dynamic navigation links from PRIMARY_NAV_LINKS. Completed by Kelly Burden - Nov 2025 */}
-      {menuOpen ? <button className="landing-basic__backdrop" aria-label="Close menu" onClick={closeMenu} /> : null}
-      <nav
-        id="landing-drawer"
-        className={"landing-basic__drawer" + (menuOpen ? " landing-basic__drawer--open" : "")}
-        aria-hidden={!menuOpen}
-      >
-        <button type="button" className="landing-basic__drawer-close" onClick={closeMenu} aria-label="Close menu">
-          ×
-        </button>
-        <ul className="landing-basic__drawer-list">
-          {PRIMARY_NAV_LINKS.map(link => (
-            <li key={link.path}>
-              <Link to={link.path} onClick={closeMenu}>
-                {link.label}
-              </Link>
-            </li>
-          ))}
-        </ul>
-      </nav>
     </div>
   );
 }
